// src/libs/api-client/model/userDtoReq.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDtoReq { 
    birthDate?: string;
    email: string;
    firstName?: string;
    id?: number;
    lastName?: string;
    password: string;
    username: string;
}