// src/libs/api-client/model/userDtoRes.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDtoRes { 
    birthDate?: string;
    email: string;
    firstName?: string;
    id?: number;
    lastName?: string;
    username: string;
}